test(rule): merge ajv defaults/realworld fixtures into one file
Combines the two generator scripts and their two testdata JSON fixtures
into a single scripts/gen-ajv-fixtures.js and ajv_fixtures.json, and
merges the corresponding Go test functions into one.
