+++ internal/rule/schema_test.go
@@ -365,7 +365,7 @@ func TestValidateFillsAllOfBranchDefaults(t *testing.T) {
 	// a default declared on an allOf branch's property is filled in, exactly
 	// as if that branch's properties were merged into this level. Confirmed
 	// against ajv@6: see allOf_multiple_branches_each_contribute_defaults in
-	// testdata/ajv_defaults_fixtures.json.
+	// testdata/ajv_fixtures.json.
 	schema, err := CompileSchema([]byte(`{
 		"type": "array",
 		"items": [
@@ -661,12 +661,20 @@ func TestValidateAcceptsJavaScriptOnlyRegexAsFormatRegexValue(t *testing.T) {
 	}
 }
 
-// ajvFixtureCase mirrors one entry of testdata/ajv_defaults_fixtures.json,
-// generated by scripts/gen-ajv-defaults-fixtures.js by running the same
-// {schema, input} pair through ajv@6 configured exactly the way ESLint
-// configures it (see eslint/lib/shared/ajv.js: useDefaults, missingRefs,
-// etc.) That file documents the shape and rationale of each case; run the
-// generator script again to regenerate it if ajv's behavior is ever in doubt.
+// ajvFixtureCase mirrors one entry of testdata/ajv_fixtures.json, generated
+// by scripts/gen-ajv-fixtures.js by running each {schema, input} pair
+// through ajv@6 configured exactly the way ESLint configures it (see
+// eslint/lib/shared/ajv.js: useDefaults, missingRefs, etc.) The corpus mixes
+// synthetic cases pinning down specific useDefaults edge cases (nested/
+// tuple/$ref gaps, allOf vs. anyOf/oneOf exclusion, additionalProperties/
+// patternProperties, etc.) with real-world cases sourced from real ESLint /
+// typescript-eslint / eslint-plugin-unicorn / eslint-plugin-react /
+// eslint-plugin-vue rule option schemas, exercising general validation
+// correctness: definitions/$defs + $ref chains, oneOf/anyOf/allOf exclusion,
+// tuple vs. list-style items, additionalItems, additionalProperties-as-
+// schema, minProperties, required, and uniqueItems. That file documents the
+// shape and rationale of each case; run the generator script again to
+// regenerate it if ajv's behavior is ever in doubt.
 type ajvFixtureCase struct {
 	Name   string          `json:"name"`
 	Schema json.RawMessage `json:"schema"`
@@ -675,50 +683,8 @@ type ajvFixtureCase struct {
 	Valid  bool            `json:"valid"`
 }
 
-func TestValidateMatchesAjvDefaultsFixtures(t *testing.T) {
-	data, err := os.ReadFile(filepath.Join("testdata", "ajv_defaults_fixtures.json"))
-	if err != nil {
-		t.Fatalf("failed to read ajv fixtures: %v", err)
-	}
-	var cases []ajvFixtureCase
-	if err := json.Unmarshal(data, &cases); err != nil {
-		t.Fatalf("failed to parse ajv fixtures: %v", err)
-	}
-	if len(cases) == 0 {
-		t.Fatal("expected at least one ajv fixture case")
-	}
-
-	for _, c := range cases {
-		t.Run(c.Name, func(t *testing.T) {
-			schema, err := CompileSchema(c.Schema)
-			if err != nil {
-				t.Fatalf("CompileSchema failed: %v", err)
-			}
-
-			got, err := schema.Validate(c.Input)
-			if gotValid := err == nil; gotValid != c.Valid {
-				t.Errorf("valid = %v, want %v (err: %v)", gotValid, c.Valid, err)
-			}
-			if want := any(c.Output); !reflect.DeepEqual(got, want) {
-				t.Errorf("got %#v, want %#v", got, want)
-			}
-		})
-	}
-}
-
-// TestValidateMatchesAjvRealWorldFixtures runs testdata/ajv_realworld_fixtures.json:
-// {schema, input} pairs sourced from real ESLint / typescript-eslint /
-// eslint-plugin-unicorn / eslint-plugin-react / eslint-plugin-vue rule
-// option schemas (generated by scripts/gen-ajv-realworld-fixtures.js), run
-// through the same ESLint-configured ajv@6 instance as the synthetic
-// defaults fixtures above. Unlike ajv_defaults_fixtures.json — which
-// targets the useDefaults corpus specifically — this corpus exercises
-// general validation correctness against schema shapes actually shipped by
-// real rules: definitions/$defs + $ref chains, oneOf/anyOf/allOf exclusion,
-// tuple vs. list-style items, additionalItems, additionalProperties-as-schema,
-// minProperties, required, and uniqueItems.
-func TestValidateMatchesAjvRealWorldFixtures(t *testing.T) {
-	data, err := os.ReadFile(filepath.Join("testdata", "ajv_realworld_fixtures.json"))
+func TestValidateMatchesAjvFixtures(t *testing.T) {
+	data, err := os.ReadFile(filepath.Join("testdata", "ajv_fixtures.json"))
 	if err != nil {
 		t.Fatalf("failed to read ajv fixtures: %v", err)
 	}
